Fix: Fixed false pass of New Feed Selection
Populating secondEntries in the beforeEach function was loading the same nodes as the first entries but was passing the test anyway because the nodes were not the same nodes. So we move the secondEntries to the it function. To solve the node issue, we grab the innerHTML to compare with the first entry.